WebIDL/tests/submissions/heycam/unsigned-long-001.html

<!DOCTYPE html>
<title>unsigned long</title>
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="utils.js"></script>
<div id="log"></div>

<script>
// 4.2.9 unsigned long
//
// Converting a JS value to an IDL unsigned long uses ECMAScript's ToUint32
// operation.

var values = [
  undefined,    0,           "undefined",
  null,         0,           "null",
  false,        0,           "false",
  true,         1,           "true",
  0,            0,           "0",
  1,            1,           "1",
  -1,           4294967295,  "-1",
  0.5,          0,           "0.5",
  Infinity,     0,           "Infinity",
  -Infinity,    0,           "-Infinity",
  NaN,          0,           "NaN",
  -0,           0,           "-0",
  "",           0,           '""',
  "1",          1,           '"1"',
  "hello",      0,           '"hello"',
  { },          0,           "{ }",
  [],           0,           "[]",
  /abc/,        0,           "/abc/",
  document,     0,           "document",
  -2147483648,  2147483648,  "-2147483648",
  -2147483649,  2147483647,  "-2147483649",
  4294967295,   4294967295,  "4294967295",
  4294967296,   0,           "4294967296",
  4294967297,   1,           "4294967297",
  -0.5,         0,           "-0.5",
  -1.5,         4294967295,  "-1.5",
  1.5,          1,           "1.5"
];

// feature: Document.createNodeIterator (operation) (with argument of type unsigned long)

for (var i = 0; i < values.length; i += 3) {
  var value = values[i];
  var expected = values[i + 1];
  var desc = values[i + 2];
  test(function() {
    var it = document.createNodeIterator(document, value, null);
    assert_equals(it.whatToShow, expected);
  }, "passing " + desc + " to Document.createNodeIterator coerces the value to an IDL unsigned long correctly");
}

for (var i = 0; i < values.length; i += 3) {
  var value = values[i];
  var expected = values[i + 1];
  var desc = values[i + 2];
  test(function() {
    var objectValue = { valueOf: function() { return value; }, toString: function() { return { } } };
    if (value === null || typeof value != "object") {
      var it = document.createNodeIterator(document, objectValue, null);
      assert_equals(it.whatToShow, expected);
    } else {
      // ToPrimitive() will throw a TypeError if valueOf and toString both return an object.
      assert_throws(new TypeError(), function() { document.createNodeIterator(document, objectValue, null); });
    }
  }, "passing { valueOf: function() { return " + desc + " } } to Document.createNodeIterator coerces the value to an IDL unsigned long correctly");
}

for (var i = 0; i < values.length; i += 3) {
  var value = values[i];
  var expected = values[i + 1];
  var desc = values[i + 2];
  test(function() {
    var objectValue = { valueOf: function() { return { } }, toString: function() { return value; } };
    if (value === null || typeof value != "object") {
      var it = document.createNodeIterator(document, objectValue, null);
      assert_equals(it.whatToShow, expected);
    } else {
      // ToPrimitive() will throw a TypeError if valueOf and toString both return an object.
      assert_throws(new TypeError(), function() { document.createNodeIterator(document, objectValue, null); });
    }
  }, "passing { toString: function() { return " + desc + " } } to Document.createNodeIterator coerces the value to an IDL unsigned long correctly");
}

test(function() {
  assert_throws({ name: "SomeError" }, function() { document.createNodeIterator(document, { toString: function() { throw { name: "SomeError" }; } }, null); });
}, "passing { toString: function() { throw ... } } to Document.createNodeIterator will propagate the exception");
</script>
